+++ src/Application.jsx
@@ -1,9 +1,9 @@
 import Nullstack from 'nullstack'
-import CreateNFT from './CreateNFT'
+import CreateNFT from './Admin/CreateNFT'
 import Dashboard from './Dashboard'
 import Home from './Home'
 import Footer from './layout/Footer'
-import MyNFTs from './MyNFTs'
+import MyNFTs from './Admin/MyNFTs'
 import Header from './layout/Header'
 import ResellNFT from './ResellNFT'
 import Explore from './Explore'
